Create Famous People Table
This is a series of tables created by me using basic information from Google. The goal of this code is to demonstrate usage of the JOIN function as well as a INNER SELECT to show the occurrence of a project twice with two different people who worked on the project within the famous_people table.