Create Example Queries for a Sales Database
This code demonstrates an example of how I would code some queries for a dataset to show total revenue, orders by location, minimum price of a distinct product, and more.

I tried a different way of formatting the code on this one to see if it might improve the readability of my code. "select" would be the furthest left portion of the query with every group after tabbed in by one space. When another group is contained within the 1st tabbed in group that group would be tabbed into a second space.

For example:

select *
     from My_Database.Sales S
     join My_Database.Accounts Acc
            on Acc.orderID = S.order_id;